ppe_compliance_monitoring_system/ppe_compliance_monitoring_ui_repo/ppe_predictor/forms.py
from django import forms


class PredictForm(forms.Form):
    file = forms.FileField(label='Please upload an image',widget=forms.ClearableFileInput(attrs={'multiple': False,'class':'form-control'}))

