ppe_compliance_monitoring_system/ppe_compliance_monitoring_ui_repo/requirements.txt
django-crispy-forms
crispy_bootstrap5